+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## [Unreleased]
-- added `delay` property to Tween
+- added `delay` property to `TweenInfo`
+- added `parallel` method to `Tweenable`
+- refactoring
 
 ## [0.1.2] - 2018-08-24
 - Remove redundant `prop-types` peer dependency